feat: restructure schemas and exceptions
