+++ src/blog/adapters/services/post.py
@@ -2,18 +2,15 @@ from typing import Any, Optional
 
 from flask import abort, g
 
-from src.blog.domain.model.post import Post, post_factory
-from src.blog.domain.ports import (
+from src.blog.domain.model.model import Post, post_factory
+from src.blog.domain.model.schemas import (
     CreatePostInputDto,
     DeletePostInputDto,
     UpdatePostInputDto,
 )
+from src.blog.domain.ports.repositories.exceptions import BlogDBOperationError
 from src.blog.domain.ports.repositories.repository import RepositoryInterface
 from src.blog.domain.ports.services.post import PostServiceInterface
-
-
-class BlogDBOperationError(Exception):
-    ...
 
 
 class PostService(PostServiceInterface):
