+++ src/blog/adapters/services/user.py
@@ -1,13 +1,10 @@
 from typing import Any, Optional
 
 from src.blog.domain.model import User, user_factory
-from src.blog.domain.ports import RegisterUserInputDto
+from src.blog.domain.model.schemas import RegisterUserInputDto
+from src.blog.domain.ports.repositories.exceptions import UserDBOperationError
 from src.blog.domain.ports.repositories.repository import RepositoryInterface
 from src.blog.domain.ports.services.user import UserServiceInterface
-
-
-class UserDBOperationError(Exception):
-    ...
 
 
 class UserService(UserServiceInterface):
