+++ src/blog/domain/ports/services/user.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from typing import Any, Optional
 
 from src.blog.domain.model import User
-from src.blog.domain.ports import RegisterUserInputDto
+from src.blog.domain.model.schemas import RegisterUserInputDto
 from src.blog.domain.ports.repositories.repository import RepositoryInterface
